requirements.txt
requests
beautifulsoup4
pyramid-jinja2